fix dictionary pattern expansion
I'm kind of mystified by this bug. It seems that, at some point in the past,
if-let did not introduce a binding in the else branch. But testing old Janet
versions reveals that this is just not the case. It has always behaved this
way, and I don't understand how these tests were ever passing.